+++ echo/tcp-echo/client/app/client.go
@@ -155,7 +155,7 @@ func (this *EchoClient) heartbeat(session getty.Session) {
 	pkg.B = echoHeartbeatRequestString
 	pkg.H.Len = (uint16)(len(pkg.B) + 1)
 
-	if err := session.WritePkg(&pkg); err != nil {
+	if err := session.WritePkg(&pkg, WritePkgTimeout); err != nil {
 		log.Warn("session.WritePkg(session{%s}, pkg{%s}) = error{%v}", session.Stat(), pkg, err)
 		session.Close()