+++ echo/tcp-echo/change_log.md
@@ -10,6 +10,15 @@
 ## develop history ##
 ---
 
+- 2018/03/14
+    > improvement
+    * add keep alive period
+    * add write package timeout
+
+- 2018/03/10
+    > improvement
+    * use getty 0.8.1
+
 - 2016/11/19
     > 1 add client/app/config.go:GettySessionParam{CompressEncoding} to test compress websocket compression extension.
     >